add failureRetry and specifiedResultRetry
