+++ consul/client/decoder.go
@@ -17,9 +17,7 @@ package client
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/cloudwego/kitex/pkg/retry"
 	"gopkg.in/yaml.v3"
-	"strings"
 )
 
 type ConfigParser interface {
@@ -38,71 +36,4 @@ func (p *defaultParser) Decode(configType ConfigType, data []byte, config *Consu
 	default:
 		return fmt.Errorf("unsupported config data type %s", configType)
 	}
-}
-
-type Config interface {
-	String() string
-}
-
-type ConsulConfig struct {
-	ClientBasicInfo   *EndpointBasicInfo       `mapstructure:"ClientBasicInfo"`
-	HostPorts         []string                 `mapstructure:"HostPorts"`
-	DestService       *string                  `mapstructure:"DestService"`
-	Protocol          *string                  `mapstructure:"Protocol"`
-	MyConfig          Config                   `mapstructure:"MyConfig"`
-	Connection        *Connection              `mapstructure:"Connection"`
-	FailureRetry      *retry.FailurePolicy     `mapstructure:"FailureRetry"`
-	ShouldResultRetry *retry.ShouldResultRetry `mapstructure:"-"`
-}
-
-func (c *ConsulConfig) String() string {
-	var builder strings.Builder
-
-	if c.ClientBasicInfo != nil {
-		builder.WriteString(fmt.Sprintf("ClientBasicInfo: %v\n", *c.ClientBasicInfo))
-	}
-
-	if c.HostPorts != nil {
-		builder.WriteString(fmt.Sprintf("HostPorts: %v\n", c.HostPorts))
-	}
-
-	if c.DestService != nil {
-		builder.WriteString(fmt.Sprintf("DestService: %v\n", *c.DestService))
-	}
-
-	if c.Protocol != nil {
-		builder.WriteString(fmt.Sprintf("Protocol: %v\n", *c.Protocol))
-	}
-
-	if c.Connection != nil {
-		builder.WriteString(fmt.Sprintf("Connection: %v\n", *c.Connection))
-	}
-
-	if c.MyConfig != nil {
-		builder.WriteString(c.MyConfig.String())
-	}
-
-	return builder.String()
-}
-
-type EndpointBasicInfo struct {
-	ServiceName string            `mapstructure:"ServiceName"`
-	Method      string            `mapstructure:"Method"`
-	Tags        map[string]string `mapstructure:"Tags"`
-}
-
-type IdleConfig struct {
-	MinIdlePerAddress int    `mapstructure:"MinIdlePerAddress"`
-	MaxIdlePerAddress int    `mapstructure:"MaxIdlePerAddress"`
-	MaxIdleGlobal     int    `mapstructure:"MaxIdleGlobal"`
-	MaxIdleTimeout    string `mapstructure:"MaxIdleTimeout"`
-}
-type MuxConnection struct {
-	ConnNum int `mapstructure:"ConnNum"`
-}
-
-type Connection struct {
-	Method         string        `mapstructure:"Method"`
-	LongConnection IdleConfig    `mapstructure:"LongConnection"`
-	MuxConnection  MuxConnection `mapstructure:"MuxConnection"`
 }
